requirements.txt
streamlit==1.36.0
llama-index==0.10.55
llama-index-core==0.10.55
llama-index-readers-file==0.1.30
llama-index-llms-openai==0.1.25
llama-index-embeddings-openai==0.1.10
llama-parse==0.4.7
PyMuPDF==1.24.7
